test(compaction): split service machinery coverage
